Marshal i64s in args and retvals
Seems to work, but putting through the tests...

The original signature (with j, not ii) is used but the dyncall
stub goes through a legalization layer so we have to send the legalized
args with the original sig. Weird. :)

Args are passed as two 32-bit words: one low word then one high word.
Retvals are returned as the low 32-bit word directly, with the high
word hiding in a global $tempRet0, retrievable through $getTempRet0